Move withCluster to a thread
