+++ src/Test/Plutip/Internal/LocalCluster/Types.hs
@@ -11,10 +11,9 @@ import BotPlutusInterface.Types (ContractState)
 import Cardano.Api (NetworkId)
 import Cardano.BM.Tracing (Trace)
 import Cardano.Launcher.Node (CardanoNodeConn)
-import Cardano.Wallet.Primitive.Types (Block, NetworkParameters)
+import Cardano.Wallet.Shelley.Launch.Cluster (RunningNode (RunningNode))
 import Control.Exception (SomeException)
 import Data.Text (Text, intercalate, pack)
-import Ouroboros.Network.NodeToClient (NodeToClientVersionData)
 import Servant.Client (BaseUrl)
 
 -- | Environment for actions that use local cluster
@@ -27,15 +26,6 @@ data ClusterEnv = ClusterEnv
     supportDir :: FilePath
   , tracer :: Trace IO Text -- not really used anywhere now
   }
-
--- | Information about a launched node.
-data RunningNode
-  = RunningNode
-      CardanoNodeConn
-      -- ^ Socket path
-      Block
-      -- ^ Genesis block
-      (NetworkParameters, NodeToClientVersionData)
 
 -- | Helper function to get socket path from
 nodeSocket :: ClusterEnv -> CardanoNodeConn
